refactor: fixing the requested changes
